+++ HappyBall/Models/final.cs
@@ -11,10 +11,10 @@ namespace HappyBall.Models
         public int Id { get; set; }
         public int Week { get; set; }
         public string TeamName { get; set; }
-        public double PropResult { get; set; }
-        public double KingResult { get; set; }
-        public double WeekTotal { get; set; }
-        public double YearTotal { get; set; }
+        public decimal PropResult { get; set; }
+        public decimal KingResult { get; set; }
+        public decimal WeekTotal { get; set; }
+        public decimal YearTotal { get; set; }
 
 
         public Final()
